Implementation of first NSMutableAttributedString Unit Tests (still WIP)

+++ Source/NSAttributedString+OHAdditions.m
@@ -126,7 +126,7 @@
                                        context:nil];
 
     // We need to ceil the returned values (see Apple doc)
-    return CGSizeMake( ceilf(bounds.size.width), ceilf(bounds.size.height) );
+    return CGSizeMake( ceil((double)bounds.size.width), ceil((double)bounds.size.height) );
 }
 
 /******************************************************************************/
